requirements.txt
pandas==2.0.0
numpy==1.24.0
mysql-connector-python==8.1.0